Added Basic Bootstrap CI
Wikipedia says "If the bootstrap distribution of an estimator is symmetric, then percentile confidence-interval are often used; such intervals are appropriate especially for median-unbiased estimators of minimum risk (with respect to an absolute loss function). Bias in the bootstrap distribution will lead to bias in the confidence-interval" (https://en.wikipedia.org/wiki/Bootstrapping_(statistics)#Bias,_asymmetry,_and_confidence_intervals). We used a median-unbiased estimator of the p-quantile, so the Percentile Bootstrap is probably better than the Basic Bootstrap. However, I added the Basic Bootstrap for comparison.